print report--incomplete

+++ App_Allocation/models.py
@@ -21,7 +21,7 @@ class Temporary_Allocation(models.Model):
     item = models.ForeignKey(Item, on_delete=models.CASCADE)  
     warehouse = models.CharField(max_length=50)  
     quantity = models.IntegerField()  
-    price = models.BigIntegerField()  
+    price = models.DecimalField(max_digits=15, decimal_places=2)  
     created_at = models.DateTimeField(auto_now_add=True)
 
     def __str__(self):
@@ -35,7 +35,7 @@ class Final_Allocation(models.Model):
     item = models.ForeignKey(Item, on_delete=models.CASCADE)  
     warehouse = models.CharField(max_length=50)  
     quantity = models.IntegerField()  
-    price = models.BigIntegerField()  
+    price = models.DecimalField(max_digits=15, decimal_places=2)  
     created_at = models.DateTimeField(auto_now_add=True)
 
     def __str__(self):
